requirements.txt
Django==3.2.25
django_imagekit==5.0.0
Pillow==9.5.0